optimzed heat_transfer

+++ src/Models/Thermal/Thermal_Factory.jl
@@ -28,7 +28,7 @@ function init_fields(datamanager::Module)
     datamanager.create_node_field("Temperature", Float64, 1)
     datamanager.create_constant_node_field("Delta Temperature", Float64, 1)
     datamanager.create_node_field("Heat Flow", Float64, 1)
-    datamanager.create_constant_node_field("Specific Volume", Float64, 1)
+    datamanager.create_constant_node_field("Specific Volume", Int64, 1)
     # if it is already initialized via mesh file no new field is created here
     datamanager.create_constant_node_field("Surface_Nodes", Bool, 1, true)
